Return a vector of banc root ids from diverse inputs

Return a vector of banc root ids from diverse inputs

Usage

banc_ids(x, integer64 = NA)

Arguments

x

A data.frame, URL or vector of ids

integer64

Whether to return ids as bit64::integer64 or character vectors. Default value of NA leaves the ids unmodified.

Value

A vector of ids

See also

Examples

banc_ids(data.frame(rootid="648518346474360770"))
#> [1] "648518346474360770"