bancr: R Client Access to the Brain And Nerve Cord (BANC) Dataset

Access to the BANC 'righty' EM dataset, focussing on support for BANC autosegmentation. Organised as a wrapper around the fafbseg package.

Details

There are several online bancr resources in addition to the built-in R help documentation. These include:

Package README including installation details and an analysis full worked example.

Function reference

Getting help

Please see below for github source code and issues links.

See also

Author

Maintainer: Alexander Bates alexander.shakeel.bates@gmail.com (ORCID)

Authors:

Gregory Jefferis jefferis@gmail.com (ORCID)