Visualize Neurons Across Multiple Drosophila Connectomic Datasets

This function constructs a Neuroglancer scene that visualizes neurons from multiple co-registered Drosophila connectomic datasets, including BANC, FAFB, hemibrain, and MANC. It allows for simultaneous visualization of corresponding neurons across these datasets.

Usage

bancsee(
  banc_ids = NULL,
  fafb_ids = NULL,
  hemibrain_ids = NULL,
  manc_ids = NULL,
  open = FALSE
)

Arguments

banc_ids

A vector of neuron IDs from the BANC dataset. Default is NULL.

fafb_ids

A vector of neuron IDs from the FAFB dataset. Default is NULL.

hemibrain_ids

A vector of neuron IDs from the hemibrain dataset. Default is NULL.

manc_ids

A vector of neuron IDs from the MANC dataset. Default is NULL.

open

Logical; if TRUE, the function will open the Neuroglancer scene in a web browser. Default is FALSE.

Value

If open = FALSE, returns a character string containing the URL for the Neuroglancer scene. If open = TRUE, opens the Neuroglancer scene in a web browser and invisibly returns the URL.

Details

The function creates a Neuroglancer scene with multiple layers, each corresponding to a different dataset:

BANC: "segmentation proofreading" layer

FAFB: "fafb v783 imported" layer

Hemibrain: "hemibrain v1.2.1 imported" and "hemibrain v1.2.1 imported, mirrored" layers

MANC: "manc v1.2.1 imported" layer

Each dataset is assigned a unique color palette to distinguish neurons from different sources:

BANC: Blue to purple spectrum

FAFB: Red spectrum

Hemibrain: Green spectrum (original) and Yellow spectrum (mirrored)

MANC: Orange spectrum

Note

This function suppresses all warnings during execution. While this ensures smooth operation, it may hide important messages. Use with caution and refer to individual function documentation if unexpected behavior occurs.

See also

Examples

if (FALSE) { # \dontrun{
# Visualize cell type DNa01 across datasets
bancsee(banc_ids = c("720575941493078142","720575941455137261"),
        fafb_ids = c("720575940644438551","720575940627787609"),
        hemibrain_ids = c("1170939344"),
        manc_ids = c("10751","10760"),
        open = TRUE)

# Get URL without opening browser
url <- bancsee(banc_ids = c("720575941493078142"),
               fafb_ids = c("720575940644438551"),
               open = FALSE)
} # }